requirements.txt
# Updated 2025-12-11 01:30 CST by ChatGPT
fastapi==0.109.0
web3>=6.0.0
fastmcp>=0.1.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
pydantic>=2.5.3
pydantic-settings>=2.5.2
python-jose[cryptography]==3.4.0
google-cloud-secret-manager==2.23.1
passlib[bcrypt]==1.7.4
email-validator>=2.2.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-cloud-pubsub==2.19.0
google-generativeai==0.3.2
google-cloud-aiplatform
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
pgvector==0.4.2
plaid-python==17.0.0
httpx==0.28.1
requests>=2.31.0
stripe>=7.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
ruff==0.8.4
mypy==1.8.0
pytest-cov==4.1.0
python-json-logger==2.0.7
pyotp>=2.9.0
webauthn>=2.0.0
user-agents
ccxt>=4.0.0
PyYAML>=6.0.1
